--- tests/preprocessing/tile_nonuniform_images_tests.py
@@ -1,342 +1,0 @@
-import copy
-import nose.tools
-import numpy as np
-import os
-import pandas as pd
-import skimage.io as sk_im_io
-from testfixtures import TempDirectory
-import unittest
-import warnings
-
-import micro_dl.preprocessing.tile_nonuniform_images as tile_images
-import micro_dl.utils.aux_utils as aux_utils
-
-
-class TestImageTilerNonUniform(unittest.TestCase):
-
-    def setUp(self):
-        """Set up a dir for tiling with flatfield"""
-
-        self.tempdir = TempDirectory()
-        self.temp_path = self.tempdir.path
-        # Start frames meta file
-        self.meta_name = 'frames_meta.csv'
-        frames_meta = aux_utils.make_dataframe()
-        self.im = 127 * np.ones((15, 11), dtype=np.uint8)
-        self.im2 = 234 * np.ones((15, 11), dtype=np.uint8)
-        self.int2str_len = 3
-        self.channel_idx = [1, 2]
-        self.pos_idx1 = 7
-        self.pos_idx2 = 8
-
-        # write pos1 with 3 time points and 5 slices
-        for z in range(5):
-            for t in range(3):
-                for c in self.channel_idx:
-                    im_name = aux_utils.get_im_name(
-                        channel_idx=c,
-                        slice_idx=z,
-                        time_idx=t,
-                        pos_idx=self.pos_idx1,
-                    )
-                    with warnings.catch_warnings():
-                        warnings.simplefilter("ignore")
-                        sk_im_io.imsave(
-                            os.path.join(self.temp_path, im_name),
-                            self.im,
-                        )
-                    frames_meta = frames_meta.append(
-                        aux_utils.parse_idx_from_name(im_name=im_name, dir_name=self.temp_path),
-                        ignore_index=True,
-                    )
-        # write pos2 with 2 time points and 3 slices
-        for z in range(3):
-            for t in range(2):
-                for c in self.channel_idx:
-                    im_name = aux_utils.get_im_name(
-                        channel_idx=c,
-                        slice_idx=z,
-                        time_idx=t,
-                        pos_idx=self.pos_idx2,
-                    )
-                    with warnings.catch_warnings():
-                        warnings.simplefilter("ignore")
-                        sk_im_io.imsave(
-                            os.path.join(self.temp_path, im_name),
-                            self.im,
-                        )
-                    frames_meta = frames_meta.append(
-                        aux_utils.parse_idx_from_name(im_name=im_name, dir_name=self.temp_path),
-                        ignore_index=True,
-                    )
-
-        # Write metadata
-        frames_meta.to_csv(os.path.join(self.temp_path, self.meta_name), sep=',')
-        # Instantiate tiler class
-        self.output_dir = os.path.join(self.temp_path, 'tile_dir')
-
-        self.tile_inst = tile_images.ImageTilerNonUniform(
-            input_dir=self.temp_path,
-            output_dir=self.output_dir,
-            tile_size=[5, 5],
-            step_size=[4, 4],
-            depths=3,
-            channel_ids=[1, 2],
-            normalize_channels=[False, True]
-        )
-
-    def tearDown(self):
-        """Tear down temporary folder and file structure"""
-
-        TempDirectory.cleanup_all()
-        nose.tools.assert_equal(os.path.isdir(self.temp_path), False)
-
-    def test_init(self):
-        """Test init"""
-        self.assertIsNone(self.tile_inst.min_fraction)
-        nose.tools.assert_equal(self.tile_inst.channel_ids, [1, 2])
-        nose.tools.assert_list_equal(list(self.tile_inst.time_ids),
-                                     [0, 1, 2])
-        nose.tools.assert_equal(self.tile_inst.flat_field_dir, None)
-        # Depth is 3 so first and last frame will not be used
-        nose.tools.assert_list_equal(list(self.tile_inst.slice_ids), [1, 2, 3])
-        np.testing.assert_array_equal(self.tile_inst.pos_ids,
-                                      np.asarray([7, 8]))
-
-        # for each tp, ch, pos check if slice_ids are the same
-        for tp_idx in range(3):
-            for ch_idx in self.channel_idx:
-                for pos_idx in [7, 8]:
-                    if ch_idx == 1:
-                        if pos_idx == 7:
-                            sl_ids = [0, 1, 2, 3, 4]
-                            np.testing.assert_array_equal(
-                                self.tile_inst.nested_id_dict[tp_idx][ch_idx][
-                                    pos_idx],
-                                np.asarray(sl_ids)
-                            )
-                        elif pos_idx == 8 and tp_idx < 2:
-                            sl_ids = [0, 1, 2]
-                            np.testing.assert_array_equal(
-                                self.tile_inst.nested_id_dict[tp_idx][ch_idx][
-                                    pos_idx],
-                                np.asarray(sl_ids)
-                            )
-
-    def test_tile_first_channel(self):
-        """Test tile_first_channel"""
-        tile_dir = self.tile_inst.get_tile_dir()
-        ch0_ids = {}
-        # get the indices for first channel
-        for tp_idx, tp_dict in self.tile_inst.nested_id_dict.items():
-            for ch_idx, ch_dict in tp_dict.items():
-                if ch_idx == 1:
-                    ch0_dict = {ch_idx: ch_dict}
-            ch0_ids[tp_idx] = ch0_dict
-
-        # get the expected meta df
-        exp_meta = []
-        for row in [0, 4, 8, 10]:
-            for col in [0, 4, 6]:
-                for z in [1, 2, 3]:
-                    for t in [0, 1, 2]:
-                        fname = aux_utils.get_im_name(
-                            channel_idx=1,
-                            slice_idx=z,
-                            time_idx=t,
-                            pos_idx=7,
-                            ext='.npy',
-                        )
-                        tile_id = '_r{}-{}_c{}-{}_sl0-3'.format(row, row+5,
-                                                               col, col+5)
-                        fname = fname.split('.')[0] + tile_id + '.npy'
-                        cur_meta = {'channel_idx': 1,
-                                    'slice_idx': z,
-                                    'time_idx': t,
-                                    'file_name': fname,
-                                    'pos_idx': 7,
-                                    'row_start': row,
-                                    'col_start': col,
-                                    'dir_name': tile_dir}
-                        exp_meta.append(cur_meta)
-                for t in [0, 1]:
-                    fname = aux_utils.get_im_name(
-                        channel_idx=1,
-                        slice_idx=1,
-                        time_idx=t,
-                        pos_idx=8,
-                        ext='.npy',
-                    )
-                    tile_id = '_r{}-{}_c{}-{}_sl0-3'.format(row, row + 5,
-                                                            col, col + 5)
-                    fname = fname.split('.')[0] + tile_id + '.npy'
-                    cur_meta = {'channel_idx': 1,
-                                'slice_idx': 1,
-                                'time_idx': t,
-                                'file_name': fname,
-                                'pos_idx': 8,
-                                'row_start': row,
-                                'col_start': col,
-                                'dir_name': tile_dir}
-                    exp_meta.append(cur_meta)
-        exp_meta_df = pd.DataFrame.from_dict(exp_meta)
-        exp_meta_df = exp_meta_df.sort_values(by=['file_name'])
-
-        ch0_meta_df = self.tile_inst.tile_first_channel(
-            channel0_ids=ch0_ids,
-            channel0_depth=3,
-        )
-        ch0_meta_df = ch0_meta_df.sort_values(by=['file_name'])
-        self.assertListEqual(list(ch0_meta_df), list(exp_meta_df))
-        # compare values of the returned and expected dfs
-        np.testing.assert_array_equal(exp_meta_df.values, ch0_meta_df.values)
-
-    def test_tile_remaining_channels(self):
-        """Test tile_remaining_channels"""
-        tile_dir = self.tile_inst.get_tile_dir()
-        # tile channel 1
-        nested_id_dict_copy = copy.deepcopy(self.tile_inst.nested_id_dict)
-        ch0_ids = {}
-        for tp_idx, tp_dict in self.tile_inst.nested_id_dict.items():
-            for ch_idx, ch_dict in tp_dict.items():
-                if ch_idx == 1:
-                    ch0_dict = {ch_idx: ch_dict}
-                    del nested_id_dict_copy[tp_idx][ch_idx]
-            ch0_ids[tp_idx] = ch0_dict
-
-        ch0_meta_df = self.tile_inst.tile_first_channel(ch0_ids, 3)
-        # tile channel 2
-        self.tile_inst.tile_remaining_channels(
-            nested_id_dict_copy,
-            tiled_ch_id=1,
-            cur_meta_df=ch0_meta_df,
-        )
-        frames_meta = aux_utils.read_meta(self.tile_inst.tile_dir)
-        # get the expected meta df which is a concat of the first channel df
-        # and the current. it does seem to retain orig index, not sure how to
-        # replace index in-place!
-        exp_meta = []
-        for row in [0, 4, 8, 10]:
-            for col in [0, 4, 6]:
-                for z in [1, 2, 3]:
-                    for t in [0, 1, 2]:
-                        for c in self.channel_idx:
-                            fname = aux_utils.get_im_name(
-                                channel_idx=c,
-                                slice_idx=z,
-                                time_idx=t,
-                                pos_idx=7,
-                                ext='.npy',
-                            )
-                            tile_id = '_r{}-{}_c{}-{}_sl0-3'.format(row, row+5,
-                                                                    col, col+5)
-                            fname = fname.split('.')[0] + tile_id + '.npy'
-                            cur_meta = {'channel_idx': c,
-                                        'slice_idx': z,
-                                        'time_idx': t,
-                                        'file_name': fname,
-                                        'pos_idx': 7,
-                                        'row_start': row,
-                                        'col_start': col,
-                                        'dir_name': tile_dir}
-                            exp_meta.append(cur_meta)
-                for t in [0, 1]:
-                    for c in self.channel_idx:
-                        fname = aux_utils.get_im_name(
-                            channel_idx=c,
-                            slice_idx=1,
-                            time_idx=t,
-                            pos_idx=8,
-                            ext='.npy',
-                        )
-                        tile_id = '_r{}-{}_c{}-{}_sl0-3'.format(row, row + 5,
-                                                                col, col + 5)
-                        fname = fname.split('.')[0] + tile_id + '.npy'
-                        cur_meta = {'channel_idx': c,
-                                    'slice_idx': 1,
-                                    'time_idx': t,
-                                    'file_name': fname,
-                                    'pos_idx': 8,
-                                    'row_start': row,
-                                    'col_start': col,
-                                    'dir_name': tile_dir}
-                        exp_meta.append(cur_meta)
-        exp_meta_df = pd.DataFrame.from_dict(exp_meta)
-        self.assertListEqual(list(frames_meta), list(exp_meta_df))
-        self.assertEqual(len(exp_meta_df), len(frames_meta))
-
-        for i in range(len(frames_meta)):
-            act_row = frames_meta.loc[i]
-            row_idx = ((exp_meta_df['channel_idx'] == act_row['channel_idx']) &
-                       (exp_meta_df['slice_idx'] == act_row['slice_idx']) &
-                       (exp_meta_df['time_idx'] == act_row['time_idx']) &
-                       (exp_meta_df['pos_idx'] == act_row['pos_idx']) &
-                       (exp_meta_df['row_start'] == act_row['row_start']) &
-                       (exp_meta_df['col_start'] == act_row['col_start']))
-            exp_row = exp_meta_df.loc[row_idx]
-            nose.tools.assert_equal(len(exp_row), 1)
-            np.testing.assert_array_equal(act_row['file_name'],
-                                          exp_row['file_name'])
-
-    def test_tile_mask_stack(self):
-        """Test tile_mask_stack"""
-
-        # create a mask
-        mask_dir = os.path.join(self.temp_path, 'mask_dir')
-        os.makedirs(mask_dir, exist_ok=True)
-        mask_images = np.zeros((15, 11, 5), dtype='bool')
-        mask_images[4:12, 4:9, 2:4] = 1
-
-        # timepoints for testing
-        mask_meta = aux_utils.make_dataframe()
-        for z in range(5):
-            for t in range(3):
-                cur_im = mask_images[:, :, z]
-                im_name = aux_utils.get_im_name(
-                    channel_idx=3,
-                    slice_idx=z,
-                    time_idx=t,
-                    pos_idx=self.pos_idx1,
-                    ext='.npy',
-                )
-                np.save(os.path.join(mask_dir, im_name), cur_im)
-                mask_meta = mask_meta.append(
-                    aux_utils.parse_idx_from_name(im_name=im_name, dir_name=mask_dir),
-                    ignore_index=True,
-                )
-        mask_meta.to_csv(os.path.join(mask_dir, 'frames_meta.csv'), sep=',')
-
-        self.tile_inst.pos_ids = [7]
-        self.tile_inst.normalize_channels = [None, None, None, False]
-        self.tile_inst.min_fraction = 0.5
-        self.tile_inst.tile_mask_stack(
-            mask_dir,
-            mask_channel=3,
-            mask_depth=3,
-        )
-        nose.tools.assert_equal(self.tile_inst.mask_depth, 3)
-
-        frames_meta = aux_utils.read_meta(self.tile_inst.tile_dir)
-        # only 4 tiles have >= min_fraction. 4 tiles x 3 slices x 3 tps
-        self.assertEqual(len(frames_meta), 36)
-        exp_cols = ['channel_idx',
-                    'slice_idx',
-                    'time_idx',
-                    'file_name',
-                    'pos_idx',
-                    'row_start',
-                    'col_start',
-                    'dir_name']
-        self.assertListEqual(list(frames_meta), exp_cols)
-        self.assertListEqual(
-            frames_meta['row_start'].unique().tolist(),
-            [4, 8],
-        )
-        self.assertListEqual(frames_meta['col_start'].unique().tolist(), [4])
-        self.assertListEqual(
-            frames_meta['slice_idx'].unique().tolist(),
-            [2, 3],
-        )
-        self.assertSetEqual(set(frames_meta.channel_idx.tolist()), {1, 2, 3})
-        self.assertSetEqual(set(frames_meta.time_idx.tolist()), {0, 1, 2})
-        self.assertSetEqual(set(frames_meta.pos_idx.tolist()), {self.pos_idx1})
